feat: drop support for controlled routing
