+++ crates/macros/src/common/mod.rs
@@ -1,8 +1,5 @@
 pub mod check;
-pub mod controller;
-pub mod http_code;
+pub mod handler;
 pub mod inject;
 pub mod injectable;
-pub mod route;
-pub mod routes;
 pub mod state;
