+++ crates/macros/src/core/mod.rs
@@ -1,4 +1,3 @@
 pub mod dto;
-pub mod module;
 pub mod param;
 pub mod query;
